% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestors.R
\name{ancestors}
\alias{ancestors}
\title{Get the ancestors of a node}
\usage{
ancestors(id, tree, n = Inf)
}
\arguments{
\item{id}{numeric ids of nodes}

\item{tree}{tree data.frame, with columns `id` and `parent_id` at least}

\item{n}{number of levels to look up; n=1 gives the parent, n=2 gives the grand-parent, etc.}
}
\value{
A vector of ids of the ancestors of the focus node(s).
}
\description{
Get the ancestors of a node
}
\examples{
tree <- data.frame(id=c(1, 2, 3, 4), parent_id=c(NA, 1, 2, 2))
ancestors(3, tree)
ancestors(3, tree, n=1)
ancestors(c(1,4), tree)
}
