% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestors.R
\name{children}
\alias{children}
\title{Get the children of a node}
\usage{
children(id, tree, n = Inf)
}
\arguments{
\item{id}{numeric ids of nodes}

\item{tree}{tree data.frame, with columns `id` and `parent_id` at least}

\item{n}{number of levels to look down; n=1 gives the direct children, n=2 gives grand children (i.e. children of all children), etc.}
}
\value{
A vector of ids of the children of the focus node(s).
}
\description{
Get the children of a node
}
\examples{
tree <- data.frame(id=c(1, 2, 3, 4), parent_id=c(NA, 1, 2, 2))
children(1, tree)
children(1, tree, n=1)
children(c(2,3), tree)
}
